requirements_txt.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
drf-yasg>=1.21.0
django-environ>=0.10.0
mysqlclient>=2.2.0
celery>=5.3.0
redis>=4.5.0
Pillow>=10.0.0
python-decouple>=3.8
django-filter>=23.0.0
django-extensions>=3.2.0
gunicorn>=21.0.0
whitenoise>=6.5.0
django-debug-toolbar>=4.1.0